+++ client/packages/service-message/index.ts
@@ -1,4 +1,4 @@
 export * from './components/ServiceMessage';
 export * from './components';
 export * from './provider/ServiceMessageProvider';
-export * from './query/use-service-message';
+export * from './hooks/use-service-message';
